+++ pkg/analyzer/visitors.go
@@ -419,6 +419,7 @@ func (visitors *Visitors) totalSessionDuration(filter *Filter) (int, error) {
 			SELECT sum(duration_seconds*sign) duration_seconds
 			FROM session t `)
 
+	// TODO tags?
 	if len(filter.Path) != 0 || len(filter.PathPattern) != 0 {
 		q.from = pageViews
 		whereTime := q.whereTime()