+++ pkg/tracker/event_test.go
@@ -22,11 +22,13 @@ func TestEventOptions_getMetaData(t *testing.T) {
 			"":        "ignore",
 		},
 	}
-	k, v := options.getMetaData()
-	assert.Len(t, k, 2)
-	assert.Len(t, v, 2)
+	k, v := options.getMetaData([]string{"author", "key"}, []string{"John", "override"})
+	assert.Len(t, k, 3)
+	assert.Len(t, v, 3)
+	assert.Contains(t, k, "author")
 	assert.Contains(t, k, "key")
 	assert.Contains(t, k, "hello")
+	assert.Contains(t, v, "John")
 	assert.Contains(t, v, "value")
 	assert.Contains(t, v, "world")
 }