+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 6.6.7
+
+* improved User-Agent filter
+
 ## 6.6.6
 
 * improved User-Agent filter